fix: unbinding many events with same function ref
Before this change adding multiple event names that share the same
function reference would cause later events to not be unbound - because
the callback was deleted from the callbacks map. This change removes the
deletion of those callbacks -keeping them in memory just incase any
other event listeners are bound to that function reference. It also
removes the optimisation case where if a callback map is empty then it
is removed from the element map - as this will no longer ever be true
(all callbacks are kept).