feat: delegate to standard calls if possible
If addEventListener is called with no options - or options is a simple
object, then we don't need to add it to the element map to store. As
soon as that element has an eventListener with passive/once it can start
populating the map. Remove event listener can then just blindly try to
call the originalRemoveEventListener to remove any simple event
listeners. This optimisation means that for event listeners not using
passive/once options, they fall down a happy path of simply using native
events listeners.